+++ code/crates/app/src/spawn.rs
@@ -1,9 +1,10 @@
 //! Utility functions for spawning the actor system and connecting it to the application.
 
 use std::path::Path;
+use std::sync::Arc;
 use std::time::Duration;
 
-use eyre::Result;
+use eyre::{eyre, Result};
 use tokio::task::JoinHandle;
 use tracing::Span;
 
@@ -11,8 +12,9 @@ use malachitebft_engine::consensus::{Consensus, ConsensusCodec, ConsensusParams,
 use malachitebft_engine::host::HostRef;
 use malachitebft_engine::network::{Network, NetworkRef};
 use malachitebft_engine::node::{Node, NodeRef};
-use malachitebft_engine::sync::{Params as SyncParams, Sync, SyncCodec, SyncRef};
+use malachitebft_engine::sync::{Params as SyncParams, Sync, SyncCodec, SyncMsg, SyncRef};
 use malachitebft_engine::util::events::TxEvent;
+use malachitebft_engine::util::output_port::OutputPort;
 use malachitebft_engine::wal::{Wal, WalCodec, WalRef};
 use malachitebft_network::{
     ChannelNames, Config as NetworkConfig, DiscoveryConfig, GossipSubConfig, NetworkIdentity,
@@ -80,7 +82,7 @@ pub async fn spawn_consensus_actor<Ctx>(
     network: NetworkRef<Ctx>,
     host: HostRef<Ctx>,
     wal: WalRef<Ctx>,
-    sync: Option<SyncRef<Ctx>>,
+    sync: Arc<OutputPort<SyncMsg<Ctx>>>,
     metrics: Metrics,
     tx_event: TxEvent<Ctx>,
 ) -> Result<ConsensusRef<Ctx>>
@@ -153,6 +155,7 @@ pub async fn spawn_sync_actor<Ctx, Codec>(
     ctx: Ctx,
     network: NetworkRef<Ctx>,
     host: HostRef<Ctx>,
+    consensus: ConsensusRef<Ctx>,
     sync_codec: Codec,
     config: &ValueSyncConfig,
     registry: &SharedRegistry,
@@ -163,6 +166,10 @@ where
 {
     if !config.enabled {
         return Ok(None);
+    }
+
+    if config.enabled && config.batch_size == 0 {
+        return Err(eyre!("Value sync batch size cannot be zero"));
     }
 
     let params = SyncParams {
@@ -192,6 +199,7 @@ where
         ctx,
         network,
         host,
+        consensus,
         params,
         sync_codec,
         sync_config,