+++ code/crates/core-consensus/src/handle.rs
@@ -1,5 +1,6 @@
 mod decide;
 mod driver;
+mod finalize;
 mod liveness;
 mod proposal;
 mod propose;
@@ -46,8 +47,17 @@ where
     Ctx: Context,
 {
     match input {
-        Input::StartHeight(height, validator_set, is_restart) => {
-            reset_and_start_height(co, state, metrics, height, validator_set, is_restart).await
+        Input::StartHeight(height, validator_set, is_restart, target_time) => {
+            reset_and_start_height(
+                co,
+                state,
+                metrics,
+                height,
+                validator_set,
+                is_restart,
+                target_time,
+            )
+            .await
         }
         Input::Vote(vote) => on_vote(co, state, metrics, vote).await,
         Input::Proposal(proposal) => on_proposal(co, state, metrics, proposal).await,