+++ code/crates/engine/src/util/events.rs
@@ -6,8 +6,8 @@ use derive_where::derive_where;
 use tokio::sync::broadcast;
 
 use malachitebft_core_consensus::{
-    Error as ConsensusError, LocallyProposedValue, ProposedValue, Role, SignedConsensusMsg,
-    WalEntry,
+    Error as ConsensusError, LocallyProposedValue, MisbehaviorEvidence, ProposedValue, Role,
+    SignedConsensusMsg, WalEntry,
 };
 use malachitebft_core_types::{
     CommitCertificate, Context, PolkaCertificate, Round, RoundCertificate, SignedVote, ValueOrigin,
@@ -51,7 +51,13 @@ pub enum Event<Ctx: Context> {
     Received(SignedConsensusMsg<Ctx>),
     ProposedValue(LocallyProposedValue<Ctx>),
     ReceivedProposedValue(ProposedValue<Ctx>, ValueOrigin),
-    Decided(CommitCertificate<Ctx>),
+    Decided {
+        commit_certificate: CommitCertificate<Ctx>,
+    },
+    Finalized {
+        commit_certificate: CommitCertificate<Ctx>,
+        evidence: MisbehaviorEvidence<Ctx>,
+    },
     RepublishVote(SignedVote<Ctx>),
     RebroadcastRoundCertificate(RoundCertificate<Ctx>),
     SkipRoundCertificate(RoundCertificate<Ctx>),
@@ -82,7 +88,35 @@ impl<Ctx: Context> fmt::Display for Event<Ctx> {
                     "ReceivedProposedValue(value: {value:?}, origin: {origin:?})"
                 )
             }
-            Event::Decided(cert) => write!(f, "Decided(value: {})", cert.value_id),
+            Event::Decided { commit_certificate } => {
+                write!(
+                    f,
+                    "Decided(value: {}, signatures: {})",
+                    commit_certificate.value_id,
+                    commit_certificate.commit_signatures.len()
+                )
+            }
+            Event::Finalized {
+                commit_certificate,
+                evidence,
+            } => {
+                if evidence.is_empty() {
+                    write!(
+                        f,
+                        "Finalized(value: {}, signatures: {})",
+                        commit_certificate.value_id,
+                        commit_certificate.commit_signatures.len()
+                    )
+                } else {
+                    write!(
+                        f,
+                        "Finalized(value: {}, signatures: {}, evidence: {:?})",
+                        commit_certificate.value_id,
+                        commit_certificate.commit_signatures.len(),
+                        evidence
+                    )
+                }
+            }
             Event::RepublishVote(vote) => write!(f, "RepublishVote(vote: {vote:?})"),
             Event::RebroadcastRoundCertificate(certificate) => write!(
                 f,