+++ code/crates/test/app/src/state.rs
@@ -291,8 +291,35 @@ impl State {
         self.store.get_decided_value(height).await.ok().flatten()
     }
 
-    /// Commits a value with the given certificate, updating internal state
-    /// and moving to the next height
+    /// Stores a value with the given certificate without updating internal state or moving to the next height.
+    pub async fn store_decided(
+        &mut self,
+        certificate: CommitCertificate<TestContext>,
+    ) -> eyre::Result<()> {
+        let (height, round, value_id) =
+            (certificate.height, certificate.round, certificate.value_id);
+
+        // Get the first proposal with the given value id. There may be multiple identical ones
+        // if peers have restreamed at different rounds.
+        let Ok(Some(proposal)) = self
+            .store
+            .get_undecided_proposal_by_value_id(value_id)
+            .await
+        else {
+            return Err(eyre!(
+                "Trying to store decided value with value id {value_id} at height {height} and round {round} for which there is no proposal"
+            ));
+        };
+
+        self.store
+            .store_decided_value(&certificate, proposal.value)
+            .await?;
+
+        Ok(())
+    }
+
+    /// Commits (or updates) a value with the given certificate, updating internal state
+    /// and moving to the next height.
     pub async fn commit(
         &mut self,
         certificate: CommitCertificate<TestContext>,
@@ -571,7 +598,7 @@ fn factor_value(value: Value) -> Vec<u64> {
 
     let mut i = 2;
     while i * i <= n {
-        if n % i == 0 {
+        if n.is_multiple_of(i) {
             factors.push(i);
             n /= i;
         } else {