+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### `app-channel`
+- Add builder pattern for custom actor injection
 - Make consensus request channel capacity configurable
 - Refactor infrastructure for spawning a channel-based application
 
@@ -43,7 +44,9 @@
 - Remove signing of proposal parts
 
 ### `sync`
+- Fix partial range request not being tracked in pending requests
 - Initial random (fixed) period adjustment in sync status ticker
+- Refactor sync actor to notify consensus of sync responses
 - Support batch retrieval of decided values
 - Validate value request ranges before processing